location: take unreliable updates into account
The idea is that if we constantly receive number of anchors
less than 4 and these anchors are all the same (anchors in
the list are the same) we consider DWM update as unreliable.

Kalman behaves well if we get even 1 anchor, but on the
next update iteration we should get another anchor. Obviously
we can't get a good prediction/estimation of the position
if we get always the same anchors.

Here in this patch I try to compare list of current anchors
and the list from previous anchors. If all are the same we
can handle only 3 updates, then localization stops.