droneloc: override loc['ts'] with local time
We have two different types of packages: from DWM and from parrot.
These packages have different timestamps, so that it is impossible
to calculate 'dt' by subtracting these timestamps or dt goes
negative. Also these packages can be handled in different order,
so again 'dt' can go negative.

In order to prevent negative 'dt', override 'ts' with the local
timestamp just on 'dt' calculation.

I hope this is a temporal solution, but for now it seems to me
the most easiest and correct way of calculating 'dt'.